Completed part 2 of day 8
Added a method to do the encoding of the string and then the characters
are counted from the returned result. +2 is added to include the actual
quotation marks that would enclose it according to the spec given in
advent of code. I should probably just enclose it within double quotes
in the first place